feat: complete week-02 assignments with detailed comments
