+++ weeks/week-02/1114405019/R11-slice.py
@@ -1,0 +1,14 @@
+"""命名切片
+
+1. 使用 slice 命名固定欄位範圍
+2. 方便從固定寬度欄位字串擷取資料
+3. 提高程式可讀性與維護性
+"""
+
+# 以 slice 物件定義欄位範圍，讓讀取固定欄位更具描述性
+record = '....................100 .......513.25 ..........'
+SHARES = slice(20, 23)
+PRICE = slice(31, 37)
+
+# 直接以命名切片從字串取值，並轉成數值後計算成本
+cost = int(record[SHARES]) * float(record[PRICE])
