--- weeks/week-02/bloom-examples/R07-ordered-dict.py
@@ -1,8 +1,0 @@
-# R7. OrderedDict（1.7）
-
-from collections import OrderedDict
-import json
-
-d = OrderedDict()
-d['foo'] = 1; d['bar'] = 2
-json.dumps(d)
