--- weeks/week-02/bloom-examples/R11-slice.py
@@ -1,6 +1,0 @@
-# R11. 命名切片 slice（1.11）
-
-record = '....................100 .......513.25 ..........'
-SHARES = slice(20, 23)
-PRICE = slice(31, 37)
-cost = int(record[SHARES]) * float(record[PRICE])
